epoll: use freezable blocking call Date	Wed, 1 May 2013 18:35:03 -0700
Avoid waking up every thread sleeping in an epoll_wait call during
suspend and resume by calling a freezable blocking call.  Previous
patches modified the freezer to avoid sending wakeups to threads
that are blocked in freezable blocking calls.

This call was selected to be converted to a freezable call because
it doesn't hold any locks or release any resources when interrupted
that might be needed by another freezing task or a kernel driver
during suspend, and is a common site where idle userspace tasks are
blocked.

Signed-off-by: Colin Cross <ccross@android.com>
Signed-off-by: flar2 <asegaert@gmail.com>

https://gitgud.io/dankzegriefer/m8-GPE-5.0.1/commit/03b7da9db90b16e556c33a9935b71b29263e6691