select: use freezable blocking call Date	Wed, 1 May 2013 18:35:04 -0700
Avoid waking up every thread sleeping in a select call during
suspend and resume by calling a freezable blocking call.  Previous
patches modified the freezer to avoid sending wakeups to threads
that are blocked in freezable blocking calls.

This call was selected to be converted to a freezable call because
it doesn't hold any locks or release any resources when interrupted
that might be needed by another freezing task or a kernel driver
during suspend, and is a common site where idle userspace tasks are
blocked.

Signed-off-by: Colin Cross <ccross@android.com>
Signed-off-by: flar2 <asegaert@gmail.com>

https://gitgud.io/dankzegriefer/m8-GPE-5.0.1/commit/3e3176f9fdf9c73e95ed5faffef76cfa6a5292f9